Add seed value to FakeRandom constructor
This allows for setting the initial seed value used by the pseudo-
random generator.  Before we were just looping over an array of
values, starting at the first and continuing to the end and then
starting over.  The seed can be used to specify a different starting
point in the array.